third_party/WebKit/LayoutTests/animations/animationworklet/multiple-effects-on-same-target-driven-by-individual-local-time-expected.txt

CONSOLE MESSAGE: line 62: background-color for the first target is: rgb(0, 255, 0)
CONSOLE MESSAGE: line 64: width for the first target is: 150px
